requirements.txt
# Flask and Server
flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
flask-cors==4.0.0

# Machine Learning Core
# Pinned to specific versions for stable, reproducible builds.
# Using torch without CUDA for CPU-only inference on Render.
torch==2.1.2
transformers==4.35.2
sentencepiece==0.1.99

# Other dependencies from original file
requests==2.31.0
Pillow==10.1.0
pdf2image==1.16.3
python-docx==1.1.0
PyMuPDF==1.23.7
reportlab==4.0.8
openai==1.3.7
httpx==0.25.2